add and comment code

+++ Application/Commands/AuthorCommands/DeleteAuthorCommand.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Shared;
+using Microsoft.EntityFrameworkCore;
 using Shared;
 
 namespace Application.Commands.AuthorCommands;
@@ -15,6 +16,11 @@ public class DeleteAuthorCommand : Command<bool>
         var author = await _authorRepository.GetByIdAsync(Id);
         if (author == null)
             return await Fail<bool>("ავტორი ვერ მოიძებნა");
+
+        var booksForAuthor = await applicationDbContext.BookAuthors.Where(x => x.AuthorId == Id && x.IsDeleted == false).Select(x => x.ProductId).ToListAsync();
+
+        if (booksForAuthor.IsNotNull())
+            return await Fail<bool>($"ავტორის წაშლა შეუძლებელია გთხოთ წაშალოთ ავტორზე მიბმული წიგნები  {string.Join(", ", booksForAuthor)} ");
 
         var result = await _authorRepository.DeleteAsync(Id);
         if (!result.Success)
